pyproject.toml
[project]
name = "chatur-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.4.0",
    "ddgs>=0.1.0",
    "edge-tts>=6.1.9",
    "fastapi>=0.110.0",
    "gtts>=2.3.0",
    "httpx>=0.27.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-groq>=0.1.0",
    "mysql-connector-python>=8.3.0",
    "pydantic>=2.0.0",
    "pymupdf>=1.23.0",
    "python-docx>=1.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "requests>=2.31.0",
    "sentence-transformers>=2.2.2",
    "speechrecognition>=3.10.0",
    "supabase>=2.0.0",
    "twilio>=8.0.0",
    "uvicorn[standard]>=0.27.0",
]
